+++ open_bus_stride_api/routers/__init__.py
@@ -10,7 +10,7 @@ ROUTER_NAMES = [
     'siri_ride_stops',
     'siri_vehicle_locations',
     'siri_snapshots',
-    'siri_time_aggregation',
+    'siri_velocity_aggregation',
 
     'gtfs_routes',
     'gtfs_stops',
